+++ all-is-cubes/src/block/builder.rs
@@ -440,15 +440,6 @@ impl From<Rgba> for Builder<'_, Atom, ()> {
     }
 }
 /// Equivalent to `Block::builder().color(color.with_alpha_one())`.
-impl From<Rgb> for Builder<'_, Atom, ()> {
-    // TODO(rgb01): Remove this impl and replace its uses with `Rgb01`, so we only use
-    // `Rgb` for emissive colors
-
-    fn from(color: Rgb) -> Self {
-        Block::builder().color(color.with_alpha_one())
-    }
-}
-/// Equivalent to `Block::builder().color(color.with_alpha_one())`.
 impl From<Rgb01> for Builder<'_, Atom, ()> {
     fn from(color: Rgb01) -> Self {
         Block::builder().color(color.with_alpha_one())